Moving set object path utililty function to separate file under utils for generic user
